simplify code

+++ srcs/math/matrix.c
@@ -6,7 +6,7 @@
 /*   By: psegura- <psegura-@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/01 20:14:53 by psegura-          #+#    #+#             */
-/*   Updated: 2024/12/25 20:54:35 by psegura-         ###   ########.fr       */
+/*   Updated: 2024/12/26 02:43:15 by psegura-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -81,9 +81,9 @@ t_point	mul_mat(const double m[3][3], t_point p)
 	t_point	result;
 
 	ft_bzero(&result, sizeof(t_point));
-	result.x = (m[0][0] * (double)p.x) + (m[0][1] * (double)p.y) + (m[0][2] * (double)p.z);
-	result.y = (m[1][0] * (double)p.x) + (m[1][1] * (double)p.y) + (m[1][2] * (double)p.z);
-	result.z = (m[2][0] * (double)p.x) + (m[2][1] * (double)p.y) + (m[2][2] * (double)p.z);
+	result.x = (m[0][0] * p.x) + (m[0][1] * p.y) + (m[0][2] * p.z);
+	result.y = (m[1][0] * p.x) + (m[1][1] * p.y) + (m[1][2] * p.z);
+	result.z = (m[2][0] * p.x) + (m[2][1] * p.y) + (m[2][2] * p.z);
 	result.c.color = p.c.color;
 	return (result);
 }
